requirements.txt
asgiref==3.3.4
coverage==5.5
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
